Move config handling to a separate module
